+++ app/policies/grade_entry_form_policy.rb
@@ -4,6 +4,10 @@ class GradeEntryFormPolicy < ApplicationPolicy
   alias_rule :grades?, :view_summary?, :update_grade?, :get_mark_columns?,
              :populate_grades_table?, :download?, :upload?, to: :grade?
 
+  def switch?
+    true
+  end
+
   def grade?
     user.admin? || user.ta?
   end