+++ db/migrate/20210730132238_remove_grouping_starter_file_timestamp.rb
@@ -1,0 +1,5 @@
+class RemoveGroupingStarterFileTimestamp < ActiveRecord::Migration[6.1]
+  def change
+    remove_column :groupings, :starter_file_timestamp
+  end
+end